fixed casting to pointer, Im not 100% sure it's currect so please give it a look
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@497 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/State.cpp
@@ -91,7 +91,7 @@ void SaveStateCallback(u64 userdata, int cyclesLate)
 	u8 *ptr = 0;
 	PointerWrap p(&ptr, PointerWrap::MODE_MEASURE);
 	DoState(p);
-	int sz = (int)(u64)ptr;
+	u64 sz = (u64)ptr;
 	u8 *buffer = new u8[sz];
 	ptr = buffer;
 	p.SetMode(PointerWrap::MODE_WRITE);
